Stoul の返り値を修正 (#26)

+++ srcs/utils/string.cpp
@@ -32,10 +32,10 @@ int Stoi(const std::string &str, std::size_t *idx, int base) {
   return static_cast<int>(num);
 }
 
-int Stoul(const std::string &str, std::size_t *idx, int base) {
+unsigned long Stoul(const std::string &str, std::size_t *idx, int base) {
   char *end;
   const char *p = str.c_str();
-  long num = std::strtoul(p, &end, base);
+  unsigned long num = std::strtoul(p, &end, base);
   if (p == end) {
     throw std::invalid_argument("Stoul");
   }
@@ -45,7 +45,7 @@ int Stoul(const std::string &str, std::size_t *idx, int base) {
   if (idx != NULL) {
     *idx = static_cast<std::size_t>(end - p);
   }
-  return static_cast<int>(num);
+  return num;
 }
 
 std::vector<std::string> SplitString(const std::string &str,
